refactor: shift to common file
