+++ bigtree/tree/construct/dictionaries.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 from typing import Any, List, Mapping, Optional, Type, TypeVar
 
+import bigtree.utils.common
 from bigtree.node import node
 from bigtree.tree.construct.strings import add_path_to_tree
 from bigtree.utils import assertions
@@ -127,7 +128,7 @@ def add_dict_to_tree_by_name(tree: T, name_attrs: Mapping[str, Mapping[str, Any]
     attr_dict_names = set(name_attrs.keys())
 
     for _node in findall(tree, lambda _node1: _node1.node_name in attr_dict_names):
-        node_attrs = assertions.filter_attributes(
+        node_attrs = bigtree.utils.common.filter_attributes(
             name_attrs[_node.node_name], omit_keys=["name"], omit_null_values=False
         )
         _node.set_attrs(node_attrs)
@@ -201,7 +202,7 @@ def dict_to_tree(
         or path_attrs.get(root_name + sep, {})
         or path_attrs.get(sep + root_name + sep, {})
     )
-    root_node_attrs = assertions.filter_attributes(
+    root_node_attrs = bigtree.utils.common.filter_attributes(
         root_node_attrs, omit_keys=["name"], omit_null_values=False
     )
     root_node = node_type(
@@ -212,7 +213,7 @@ def dict_to_tree(
 
     # Convert dictionary to dataframe
     for node_path, node_attrs in path_attrs.items():
-        node_attrs = assertions.filter_attributes(
+        node_attrs = bigtree.utils.common.filter_attributes(
             node_attrs, omit_keys=["name"], omit_null_values=False
         )
         add_path_to_tree(
