+++ bigtree/tree/export/dictionaries.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 from typing import Any, Dict, List, Optional, TypeVar
 
 from bigtree.node import node
+from bigtree.utils import common
 
 __all__ = [
     "tree_to_dict",
@@ -11,34 +12,6 @@ __all__ = [
 ]
 
 T = TypeVar("T", bound=node.Node)
-
-
-def _assemble_attributes(
-    _node: T,
-    attr_dict: Optional[Dict[str, str]],
-    all_attrs: bool,
-    data_child: Dict[str, Any] = None,
-) -> Dict[str, Any]:
-    """Assemble attributes of node into a dictionary.
-
-    Args:
-        _node: node
-        attr_dict: node attributes mapped to dictionary key, key: node attributes, value: corresponding dictionary key
-        all_attrs: indicator whether to retrieve all ``Node`` attributes, overrides `attr_dict`
-        data_child: existing attributes, if any
-
-    Returns:
-        node attributes
-    """
-    data_child = data_child or {}
-    if all_attrs:
-        data_child.update(
-            dict(_node.describe(exclude_attributes=["name"], exclude_prefix="_"))
-        )
-    elif attr_dict:
-        for k, v in attr_dict.items():
-            data_child[v] = _node.get_attr(k)
-    return data_child
 
 
 def tree_to_dict(
@@ -107,7 +80,7 @@ def tree_to_dict(
                     if _node.parent:
                         parent_name = _node.parent.node_name
                     data_child[parent_key] = parent_name
-                data_child = _assemble_attributes(
+                data_child = common.assemble_attributes(
                     _node, attr_dict, all_attrs, data_child
                 )
                 data_dict[_node.path_name] = data_child
@@ -165,7 +138,7 @@ def tree_to_nested_dict(
         if _node:
             if not max_depth or _node.depth <= max_depth:
                 data_child = {name_key: _node.node_name}
-                data_child = _assemble_attributes(
+                data_child = common.assemble_attributes(
                     _node, attr_dict, all_attrs, data_child
                 )
                 if child_key in parent_dict:
@@ -224,7 +197,7 @@ def tree_to_nested_dict_key(
         """
         if _node:
             if not max_depth or _node.depth <= max_depth:
-                data_child = _assemble_attributes(_node, attr_dict, all_attrs)
+                data_child = common.assemble_attributes(_node, attr_dict, all_attrs)
                 if child_key in parent_dict:
                     parent_dict[child_key][_node.node_name] = data_child
                 else:
